apps/worker/requirements.txt
# ContractLens Worker - populated in Phase 4
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymongo>=4.10.0
chromadb>=0.5.0
langchain>=0.3.0
langchain-text-splitters>=0.3.0
langchain-openai>=0.2.0
pypdf>=5.0.0
pdfminer.six>=20231228
httpx>=0.27.0
python-multipart>=0.0.12
beautifulsoup4>=4.12.0
